WIP refactor!: hide system_control
