+++ tests/amplitude/test_parity_prefactor.py
@@ -7,7 +7,7 @@ from expertsystem.state.particle import (
     InteractionQuantumNumberNames,
     get_interaction_property,
 )
-from expertsystem.ui.system_control import (
+from expertsystem.ui._system_control import (
     InteractionTypes,
     StateTransitionManager,
 )
