+++ tests/amplitude/test_yaml_canonical.py
@@ -7,7 +7,7 @@ import pytest
 import yaml
 
 from expertsystem.amplitude.canonical_decay import CanonicalAmplitudeGenerator
-from expertsystem.ui.system_control import (
+from expertsystem.ui._system_control import (
     InteractionTypes,
     StateTransitionManager,
 )
