+++ tests/channels/test_nbody_reactions.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from expertsystem.ui.system_control import (
+from expertsystem.ui._system_control import (
     InteractionTypes,
     StateTransitionManager,
 )
