+++ tests/channels/test_y_to_d0d0barpi0pi0.py
@@ -15,7 +15,7 @@ from expertsystem.state.particle import (
 from expertsystem.ui._default_settings import (
     create_default_interaction_settings,
 )
-from expertsystem.ui.system_control import (
+from expertsystem.ui._system_control import (
     InteractionTypes,
     StateTransitionManager,
     _change_qn_domain,
